Use raw MemoryRef inside Externals attrs
The thing is that MemRef is defined as:
```
pub struct MemoryRef(Rc<MemoryInstance>);
```

And at the same time, `MemoryInstance` looks like:
```
pub struct MemoryInstance {
    /// Memory limits.
    limits: ResizableLimits,
    /// Linear memory buffer with lazy allocation.
    buffer: RefCell<ByteBuf>,
    initial: Pages,
    current_size: Cell<usize>,
    maximum: Option<Pages>,
}
```

The structures mostly contain references and no heavy
data structures directly. And also, counting that we
have the `Rc` I think we should clone the structure,
instead of try to play sneaky with the lifetimes.

Since at the end I don't think it will make a big
difference in terms of performance.

Closes #55