Implement error handling for MemoryRef access
ExternalResolver host function implementations usually
need to access the internal `MemoryRef` of the resolver.

When this happens, we need a way to have error handling for the
cases where this memory has not been initialized/ set during the
initialization of WASM.

Therefore a new fn def has been added to the `MemoryHolder` trait
which basically checks whether the `MemoryRef` that the resolver
holds is `Some(mem)` or `None`. In the case of the last one, returs
a `Trap`.